fix(task-cuda-drain-plane-fold-every-8): fold the CUDA XYZ plane to double every 8 batches
Swap candidate B (double atomicAdd plane, -13.9%/-17.3% throughput on the
two reference boxes) for candidate E: revert 2d72f817 to bring back the
fp32 atomicAdd plane + device double fold plane + FoldDeviceXyzBatch, and
gate the simulator's fold call to every kXyzFoldEveryBatches (8) batches
instead of every batch. The fp32 chain any fold corrects for is capped at
8 batches whatever LUMICE_XYZ_DRAIN_BATCHES is, measured drift <= 0.05%
against the host double ledger (unfolded 64-batch window: +0.40%) at
-11.0%/-5.9% throughput.

The batch that ends a window skips the cadence fold: ReadbackXyzAccum runs
the same kernel in finalize mode over the residue on every drain site
(window cap and the display-cadence drains in Run()), so the tail is never
dropped and there is one owner of the drain-time fold. Comments the revert
brought back ("once per batch", "one batch long", "lives ONE drain
window") rewritten for the cadence; one task-id reference dropped.

+++ src/core/shared/accum_shared.h
@@ -3,36 +3,30 @@
 // Surface contract
 //   AccumXyzToPixel(buf, pix_flat, cmf_x, cmf_y, cmf_z, weight)
 //     Atomically adds weight * (cmf_x, cmf_y, cmf_z) to a 3-channel XYZ
-//     pixel at `pix_flat` inside a W*H*3 buffer.
+//     pixel at `pix_flat` inside a W*H*3 float buffer.
 //
 // Backend variants
 //   MSL (`__METAL_VERSION__`):   buf is `device atomic_float*`
-//   CUDA (`__CUDACC__`):         buf is `double*`       (atomicAdd intrinsic,
-//                                                       native since sm_60)
+//   CUDA (`__CUDACC__`):         buf is `float*`        (atomicAdd intrinsic)
 //   Host C++:                    buf is `float*`        (plain +=, used by
 //                                                       parity oracles; not
 //                                                       the cross-batch path)
 //
-// The Metal and host variants add into a float32 buffer that lives ONE batch
-// (Metal drains per batch) or one session (host oracle) — a bounded chain. The
-// CUDA variant is the exception because its plane is not bounded that way: the
-// third-clock drain (Simulator::kDefaultXyzDrainBatches) keeps the device plane
-// alive for a whole window of batches, and a per-exit fp32 chain that long
-// measured a window-length-dependent drift of up to +0.40% on a narrow-field
-// scene against the host's double ledger (16 batches → −0.135%, 1 → −0.008%) —
-// the pixel-level twin of the landed-weight scalar defect described in
-// cuda_trace_backend.cu's EmitToDeviceXyz. Folding the fp32 plane into a
-// double plane once per batch on device was measured correct but cost 22% of
-// throughput on a 2048×1024 plane (a full-plane pass on the stream's critical
-// path every batch), so the CUDA plane is simply double: the per-exit atomic
-// widens, nothing is added per batch, and the drain converts to float on
-// device before the D2H. The cross-batch reduction on the host side is not
-// here: RenderConsumer folds each window into a double running sum
-// (server/render.cpp), which replaced the float Neumaier pair this header used
-// to carry. Compensated float summation was measured to be no answer to an
-// unbounded chain — its compensation term is itself a float running sum and
-// drifts the same way, 4e-3 relative after 1e7 constant addends — so a new
-// long-chain accumulator takes a double, not a `NeumaierAdd`.
+// Every variant adds into a float32 buffer, and every caller bounds the chain
+// that buffer ever holds. Metal drains (reads back and zeroes) its plane every
+// batch. CUDA keeps its plane alive across a whole drain window, so the chain
+// is bounded one level down instead: the simulator has the backend fold the
+// fp32 plane into a device double plane and zero the fp32 side every
+// Simulator::kXyzFoldEveryBatches (8) batches, and the drain folds the residue
+// on the same kernel (fold_xyz_plane_kernel in cuda_trace_backend.cu) — so no
+// fp32 chain here is longer than 8 batches, whatever the window length. The
+// host oracle's buffer lives one session. The cross-batch reduction on the
+// host side is not here either: RenderConsumer folds each window into a double
+// running sum (server/render.cpp), which replaced the float Neumaier pair this
+// header used to carry. Compensated float summation was measured to be no
+// answer to an unbounded chain — its compensation term is itself a float
+// running sum and drifts the same way, 4e-3 relative after 1e7 constant addends
+// — so a new long-chain accumulator takes a double, not a `NeumaierAdd`.
 //
 // Note: the parameter type cannot be unified across all three backends because
 // MSL requires the `device atomic_float*` qualifier for atomic_fetch_add.
@@ -55,14 +49,12 @@ inline void AccumXyzToPixel(device metal::atomic_float* xyz_buf, uint32_t pix_fl
 
 #elif defined(__CUDACC__)
 // ===== CUDA =====
-__device__ inline void AccumXyzToPixel(double* xyz_buf, uint32_t pix_flat, float cmf_x, float cmf_y, float cmf_z,
+__device__ inline void AccumXyzToPixel(float* xyz_buf, uint32_t pix_flat, float cmf_x, float cmf_y, float cmf_z,
                                        float weight) {
   uint32_t base = pix_flat * 3u;
-  // The product stays a float product (same rounding as Metal / host); only the
-  // running sum is double.
-  atomicAdd(xyz_buf + base + 0u, static_cast<double>(cmf_x * weight));
-  atomicAdd(xyz_buf + base + 1u, static_cast<double>(cmf_y * weight));
-  atomicAdd(xyz_buf + base + 2u, static_cast<double>(cmf_z * weight));
+  atomicAdd(xyz_buf + base + 0u, cmf_x * weight);
+  atomicAdd(xyz_buf + base + 1u, cmf_y * weight);
+  atomicAdd(xyz_buf + base + 2u, cmf_z * weight);
 }
 
 #else
